feat(edgeless): support hollow circle and move color panel to components
