+++ packages/blocks/src/page-block/edgeless/components/component-toolbar/change-connector-button.ts
@@ -1,5 +1,5 @@
 import '../tool-icon-button.js';
-import '../../toolbar/brush-tool/color-panel.js';
+import '../color-panel.js';
 
 import { ConnectorLIcon, ConnectorXIcon } from '@blocksuite/global/config';
 import type {
@@ -18,14 +18,11 @@ import { styleMap } from 'lit/directives/style-map.js';
 import { countBy, maxBy } from '../../../../__internal__/utils/std.js';
 import type { EdgelessSelectionSlots } from '../../edgeless-page-block.js';
 import type { EdgelessSelectionState } from '../../selection-manager.js';
-import type {
-  ColorEvent,
-  EdgelessColorPanel,
-} from '../../toolbar/brush-tool/color-panel.js';
 import {
   generateConnectorPath,
   getConnectorAttachedInfo,
 } from '../../utils.js';
+import type { ColorEvent, EdgelessColorPanel } from '../color-panel.js';
 import { createButtonPopper } from '../utils.js';
 
 function getMostCommonColor(elements: ConnectorElement[]): Color | undefined {
